Impact of Bug Weed on Growth
To find out the impacts of Bug weed (Solanum mauritiana scopoli) on the growth of height, diameter, and morphological features of Cupressus lusitanica saplings in PELIS farms of XXX Forest.
Specific objectives
i.	To determine the effects of Bug weed on growth in height of Cupressus lusitanica saplings.
ii.	To estimate the effects of Bug weed on growth of diameter of Cupressus lusitanica saplings in the PELIS farms.
iii.	To establish the effects of Bug weed on morphological characteristics of Cupressus lusitanica saplings.